back-end/pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "fastapi>=0.135.2",
    "fastapi-users[sqlalchemy]>=15.0.5",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.42.0",
]
